backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-dotenv==1.0.1
pydantic==2.9.0
openai>=1.40.0
azure-ai-inference>=1.0.0b1
python-multipart>=0.0.9
rapidocr-onnxruntime>=1.3.0
Pillow>=10.0.0
